민하 : dev-minha13(스크랩 완) - develop와 충돌 확인 : 충돌 수정

+++ src/main/java/com/laonworks502/team1st/controller/users/GeneralUserController.java
@@ -42,8 +42,8 @@ public class GeneralUserController {
 //        return "generaluser/mainMypage";
 //    }
 
-
-/*    @RequestMapping(value = "/totalscrap")
+/*
+    @RequestMapping(value = "/totalscrap")
     public String totalscrap(HttpSession session,
                              Model model,
                              @ModelAttribute ScrapListBean myminiscrap100,
@@ -65,8 +65,34 @@ public class GeneralUserController {
         model.addAttribute("myminiscrap300",myminiscrap300);
 
         return "generaluser/totalscrap";
-    }*/
+    }
 
+    // 일반 회원 마이페이지
+    @RequestMapping(value = "/generalmypage")
+    public String generalmypage(HttpSession session,
+                                Model model,
+                                @ModelAttribute ScrapListBean myminiscrap100,
+                                @ModelAttribute ScrapListBean myminiscrap200,
+                                @ModelAttribute ScrapListBean myminiscrap300
+//                                @ModelAttribute GeneralUserBean gub
+                                )throws Exception{
+
+        LoginBean loginBean = (LoginBean) session.getAttribute("loginBean");
+        String email = loginBean.getEmail();
+
+        GeneralUserBean gub = gus.checkGeneraluser(email);
+
+        log.info("generalmypage:" + gub.getEmail());
+
+        model.addAttribute("gub", gub);
+
+        model.addAttribute("myminiscrap100",myminiscrap100);
+        model.addAttribute("myminiscrap200",myminiscrap200);
+        model.addAttribute("myminiscrap300",myminiscrap300);
+
+        return "generaluser/generalmypage";
+    }
+*/
 
     // 회원가입 폼
     @RequestMapping(value = "/generaluserinsert")
@@ -128,29 +154,31 @@ public class GeneralUserController {
                                      HttpServletRequest request,
                                      @RequestParam("passconfirm") String passconfirm,
                                      @RequestParam("email") String email,
-                                     @ModelAttribute GeneralUserBean gub) throws Exception {
+                                     @ModelAttribute GeneralUserBean gub,
+                                     Model model) throws Exception {
 
         GeneralUserBean old = gus.checkGeneraluser(gub.getEmail());
 
-        System.out.println(gub.getEmail());
-        System.out.println(gub.getPasswd());
-        System.out.println(gub.getSalt());
+        log.info(gub.getEmail());
+        log.info("old.getPasswd() : "+old.getPasswd());
 
         String salt = old.getSalt();
-        System.out.println(salt);
-        System.out.println("입력값의 암호화 : " + SHA256Util.getEncrypt_gu(passconfirm, salt));
+        log.info(salt);
+        log.info("탈퇴폼에서 입력한 값 암호화 : " + SHA256Util.getEncrypt_gu(passconfirm, salt));
 
         int result1 = 0;
         if(old.getPasswd().equals(SHA256Util.getEncrypt_gu(passconfirm, salt))){
 
             int result = gus.updateGeneraluser(gub);
-            if(result == 1) log.info("수정 성공");
+            if(result == 1) log.info("[" +gub.getEmail() + "] 의 정보 수정 성공");
 
-            return "generaluser/generalmypage";
+            return "redirect:/generalmypage";
 
         }else{
             result1 = 2;
             log.info("비밀번호 틀림");
+            model.addAttribute("result", result1);
+
             return "generaluser/loginResult";
         }
 
@@ -177,20 +205,42 @@ public class GeneralUserController {
     // 회원삭제 실행
     @RequestMapping("generaluserdelete_ok")
     public String generaluserdelete_ok(@RequestParam("exit_reason") String exit_reason,
+                                       @RequestParam("passconfirm") String passconfirm,
                                        HttpSession session,
-
-                                       GeneralUserBean gub) throws Exception {
+                                       GeneralUserBean gub,
+                                       Model model) throws Exception {
 
 		LoginBean loginBean = (LoginBean) session.getAttribute("loginBean");
 		String email = loginBean.getEmail();
-        
-		gub.setExit_reason(exit_reason);
-        int result = gus.deleteGeneraluser(gub);
-        if(result == 1) log.info("탈퇴사유 : " + gub.getExit_reason());
 
-        session.invalidate();
+        gub.setExit_reason(exit_reason);
 
-        return "generaluser/loginForm";
+        GeneralUserBean old = gus.checkGeneraluser(gub.getEmail());
+
+        String salt = old.getSalt();    // 바로 gub.getSalt 호출하면 null 값이 나와서 String으로 변환.
+
+        log.info(salt);
+        log.info("탈퇴폼에서 입력한 값 암호화 : " + SHA256Util.getEncrypt_gu(passconfirm, salt));
+
+        int result1 = 0;
+        if(old.getPasswd().equals(SHA256Util.getEncrypt_gu(passconfirm, salt))){
+
+            int result = gus.deleteGeneraluser(gub);
+            if(result == 1) log.info("[" + gub.getEmail() +"] 의 탈퇴사유 : " + gub.getExit_reason());
+
+            session.invalidate();
+
+            result1 = 4;
+            model.addAttribute("result",result1);
+
+        }else{
+            result1 = 2;
+            log.info("비밀번호 틀림");
+            model.addAttribute("result", result1);
+
+        }
+
+        return "generaluser/loginResult";
     }
 
     // 이력서 업로드
@@ -209,12 +259,12 @@ public class GeneralUserController {
         String fileRealName = file.getOriginalFilename(); //파일명을 얻어낼 수 있는 메서드!
         long size = file.getSize(); //파일 사이즈
 
-        System.out.println("파일명 : "  + fileRealName);
-        System.out.println("용량크기(byte) : " + size);
+        log.info("파일명 : "  + fileRealName);
+        log.info("용량크기(byte) : " + size);
 
         // 서버에 저장할 파일이름 fileextension으로 .jsp이런식의  확장자 명을 구함
         String fileExtension = fileRealName.substring(fileRealName.lastIndexOf("."),fileRealName.length());
-        System.out.println(fileExtension);
+        log.info(fileExtension);
         String uploadFolder = "C:\\test\\upload";
 
 		/* UUID 설명 :
@@ -225,17 +275,17 @@ public class GeneralUserController {
 		 */
         // 파일명 랜덤설정 함수인 UUID 설정 부분. 현재 컨트롤러에서 사용하지는 않을 것임.
         UUID uuid = UUID.randomUUID();
-        System.out.println(uuid.toString());
+        log.info(uuid.toString());
         String[] uuids = uuid.toString().split("-");
 
         String uniqueName = uuids[0];
-        System.out.println("생성된 고유문자열:" + uniqueName);
-        System.out.println("확장자명:" + fileExtension);
+        log.info("생성된 고유문자열:" + uniqueName);
+        log.info("확장자명:" + fileExtension);
 
         // path + uniqueName
         String resume = uniqueName+fileExtension;
         // uploadFolder + "\\" +
-        System.out.println("resume = "+resume);
+        log.info("resume = "+resume);
         // 파일명 랜덤설정 함수인 UUID 설정 부분
 
 
@@ -274,15 +324,15 @@ public class GeneralUserController {
         // => C:\Users\sky66\IdeaProjects\team1st\src\main\webapp 안에 upload 생성.
         // ex) String uploadFolder = request.getRealPath("upload");
 
-        System.out.println("download...");
+        log.info("download...");
 
         String path = uploadFolder + "\\" + resume;
-        System.out.println("path=" + path);
+        log.info("path=" + path);
 
         File file = new File(path);     // path = uploadFolder+"\\"+uniqueName + fileExtension;
         String downName = file.getName(); //다운로드 받을 파일명을 절대경로로  구해옴
 
-        System.out.println(downName);
+        log.info(downName);
 
         // 이 부분이 한글 파일명이 깨지는 것을 방지해 줍니다
         downName = new String(downName.getBytes("utf-8"), "iso-8859-1");
